nastar game updates
I wanted to set up an array of buttons to minimize the amount of functions to write for toggling the decor elements.

instead of having a bunch of functions, now i have a bunch of arrays. but now we can toggle the visibility of different elements. now i just need to figure out how to export that to the next scene -_-"

maybe check for visibility and then save as a union sprite?